Splitted container wrapper into separated file
